+++ src/gwt/test/org/rstudio/studio/client/workbench/commands/DummyCommands.java
@@ -319,6 +319,11 @@ public class DummyCommands extends Commands {
     public AppCommand insertChunkD3() {
         return null;
     }
+    
+    @Override
+    public AppCommand insertChunkJulia() {
+        return null;
+    }
 
     @Override
     public AppCommand insertChunkPython() {