+++ NEWS.md
@@ -45,6 +45,7 @@
 - Fixed an issue where quoted variable names were not completed properly in dplyr pipes (#15161)
 - Fixed issue with highlight of `tikz` code chunks in R Markdown documents (#15019)
 - RStudio now uses current session repositories when installing package dependencies via background jobs (#10016)
+- Fixed issue where collapsed raw chunks were displayed with an incorrect label in the Visual Editor (#14594)
 
 #### Posit Workbench
 - Fixed an issue with Workbench login not respecting "Stay signed in when browser closes" when using Single Sign-On (rstudio-pro#5392)